refactor: ParseUtils 패키지 위치 개선
